+++ src/main/java/se/puggan/factory/container/slot/HiddenSlot.java
@@ -2,8 +2,12 @@ package se.puggan.factory.container.slot;
 
 import net.minecraft.inventory.IInventory;
 
+/**
+ * A slot that automaticly hides it selfs when empty
+ */
 public class HiddenSlot extends LockedSlot {
     private final int slotIndex;
+
     public HiddenSlot(IInventory inventory, int index, int xPosition, int yPosition) {
         super(inventory, index, xPosition, yPosition, !inventory.getStackInSlot(index).isEmpty());
         slotIndex = index;