+++ src/main/java/se/puggan/factory/container/slot/ItemSlot.java
@@ -1,12 +1,14 @@
 package se.puggan.factory.container.slot;
 
-import net.minecraft.inventory.IInventory;
+import javax.annotation.Nullable;
 import net.minecraft.inventory.container.Slot;
+import net.minecraft.inventory.IInventory;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-import javax.annotation.Nullable;
-
+/**
+ * A slot for a specific item, No other items can be put here
+ */
 public class ItemSlot extends Slot {
     public Item lockedItem = null;
     public boolean enabled;
@@ -29,9 +31,11 @@ public class ItemSlot extends Slot {
     public void lockItem(Slot slot) {
         lockItem(slot.getStack().getItem());
     }
+
     public void lockItem(ItemStack stack) {
         lockItem(stack.getItem());
     }
+
     public void lockItem(@Nullable Item item) {
         lockedItem = item;
         enabled = item != null;