+++ src/main/java/se/puggan/factory/util/IntPair.java
@@ -2,6 +2,10 @@ package se.puggan.factory.util;
 
 import java.util.Collection;
 
+/**
+ * A rpw for list of ints (a) sorted by another list of ints (b)
+ * for exemple useful as `TreSet<IntPair>`
+ */
 public class IntPair implements Comparable<IntPair> {
     private final int a;
     private final int b;
@@ -19,7 +23,7 @@ public class IntPair implements Comparable<IntPair> {
     public static int[] aArray(Collection<IntPair> list) {
         int[] array = new int[list.size()];
         int index = 0;
-        for(IntPair ip : list) {
+        for (IntPair ip : list) {
             array[index++] = ip.a;
         }
         return array;